+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -13,7 +13,7 @@ import kitchenpos.application.dto.response.MenuGroupResponse;
 import kitchenpos.application.dto.response.MenuResponse;
 import kitchenpos.application.dto.response.OrderResponse;
 import kitchenpos.application.dto.response.ProductResponse;
-import kitchenpos.application.dto.response.TableResponse;
+import kitchenpos.table.service.dto.TableResponse;
 import kitchenpos.menu.service.MenuService;
 import kitchenpos.menugroup.service.MenuGroupService;
 import kitchenpos.order.service.OrderService;
@@ -24,6 +24,7 @@ import kitchenpos.supports.MenuGroupFixture;
 import kitchenpos.supports.OrderFixture;
 import kitchenpos.supports.OrderTableFixture;
 import kitchenpos.supports.ProductFixture;
+import kitchenpos.table.service.TableService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
