+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -7,12 +7,12 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import java.util.List;
 import kitchenpos.application.dto.request.OrderStatusRequest;
 import kitchenpos.application.dto.request.TableGroupRequest;
-import kitchenpos.application.dto.request.TableRequest;
+import kitchenpos.table.service.dto.TableRequest;
 import kitchenpos.application.dto.response.MenuGroupResponse;
 import kitchenpos.application.dto.response.MenuResponse;
 import kitchenpos.application.dto.response.OrderResponse;
 import kitchenpos.application.dto.response.ProductResponse;
-import kitchenpos.application.dto.response.TableResponse;
+import kitchenpos.table.service.dto.TableResponse;
 import kitchenpos.menu.service.MenuService;
 import kitchenpos.menugroup.service.MenuGroupService;
 import kitchenpos.order.service.OrderService;
@@ -24,6 +24,7 @@ import kitchenpos.supports.OrderFixture;
 import kitchenpos.supports.OrderTableFixture;
 import kitchenpos.supports.ProductFixture;
 import kitchenpos.supports.TableGroupFixture;
+import kitchenpos.table.service.TableService;
 import kitchenpos.tablegroup.service.TableGroupService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
